Create Matrix and MatrixWeightedMajority
Matrix objects represent algorithms that compute a matrix from a profile. They will be used as parameters for many rules (maximin, etc.). This will give us a lot of flexibility when choosing the variant of the rule that we want (especially about how to count unordered or absent candidates, indifference, etc.).