The converter becomes a parameter of the voting rule
Before this update, each rule had:
* A `default converter' (as a parameter of the rule, declared in the __init__),
* And an `override-converter' (as a variable in the __call__).
This seemed too complicated for the user, and I simply decided that it is part of the rule (hence, a parameter in the __init__). If you want to use the same rule with different converters (e.g. Approval applied to Range Voting ballot with different approval thresholds), then you may either:
* Define several rules, each with a different converter,
* Make several copies of the rule and change their converters,
* Use the same rule and change its converter before each call.
Anyway, this new philosophy makes more sense: in the example of Approval on RV ballots, each time you choose an approval threshold, you actually define a new voting rule (i.e. a mapping from some ballots to an outcome).