Focus on exact arithmetic
The core of this commit is Utils.convert_number and Utils.my_division. We should use convert_number each time an input is given, and use my_division each time we make a division. In both cases, the goal is to avoid floats and use exact fractions instead. Each time we output a number, it is also a good idea to use convert_number, because it will convert fractions like Fraction(3, 1) to an integer like 3.