New class structure for Priority
Now Priority gives a (possibly defective) function of comparison over sets of candidates (and not only candidates). A Priority in the old sense in now simply a Priority that knows only how to deal with singletons.

Motivation: a variety of priority setttings could be defined that are able to compare two committees only in some particular cases. In fact it was already the case in Whalrus with Priority.UNAMBIGUOUS, which can never decide between two different candidates (or committtees). This new implementation, more general, is able to deal with priority setting that can decide in some cases and not decide in others.